Update nexran.cc with TCP server
This includes the required dependencies, adds the needed functions, and calls those functions to add the functionality of communicating to the nexran xapp. Theoretically the only additional edit should be calling the TCP_server function in main to setup the connection which can be threaded to allow other function to operate while the TCP server waits for nexran to connect and make the socket. There may be some bugs because I had to convert from python to c++ but it should be mostly fine as I ran it in a build and it didnt have any errors.